requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
openstacksdk==4.1.0
keystoneauth1==5.8.0
python-dotenv==1.0.1
